backend/pyproject.toml
[project]
name = "local-llm-backend"
version = "0.1.0"
description = "FastAPI backend and MCP server for the local LLM project."
authors = [{ name = "Brendan Campbell" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.27.0,<1.0.0",
    "pydantic>=2.6.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "httpx>=0.26.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
    "httpx>=0.26.0,<1.0.0",
    "pytest>=8.0.0,<9.0.0",
    "ruff>=0.3.0,<1.0.0"
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
